Store database setting values as JSON
value is stored in a text field on purpose. MySQL is more forgiving than MSSQL when trying to insert an integer into a text field, So we're forcing everything to JSON so it's always treated as a string. Has the additional bonus in	that we can store complex values in the value column for future use (i.e, arrays for dropdowns etc)